fix: reuse eventStyleGetter-function instead of duplicating it
- eventStyleGetter() moved to components/common/calendarUtils.ts
